+++ src/main/java/cn/yuqi/mini/spring/beans/factory/BeanFactory.java
@@ -1,7 +1,0 @@
-package cn.yuqi.mini.spring.beans.factory;
-
-import cn.yuqi.mini.spring.beans.BeansException;
-
-public interface BeanFactory {
-    Object getBean(String beanName) throws BeansException;
-}