+++ src/main/java/cn/yuqi/mini/springframework/beans/factory/BeanFactory.java
@@ -1,0 +1,7 @@
+package cn.yuqi.mini.springframework.beans.factory;
+
+import cn.yuqi.mini.springframework.beans.BeansException;
+
+public interface BeanFactory {
+    Object getBean(String beanName) throws BeansException;
+}