+++ src/main/java/cn/yuqi/mini/springframework/beans/factory/config/SingletonBeanRegistry.java
@@ -1,0 +1,8 @@
+package cn.yuqi.mini.springframework.beans.factory.config;
+
+//负责 管理已经创建完成的单例 Bean 实例
+public interface SingletonBeanRegistry {
+    void registerSingleton(String beanName, Object bean);
+    Object getSingleton(String beanName);
+    boolean containsSingleton(String beanName);
+}