+++ src/main/java/cn/yuqi/mini/springframework/beans/factory/support/BeanDefinitionRegistry.java
@@ -1,0 +1,9 @@
+package cn.yuqi.mini.springframework.beans.factory.support;
+
+import cn.yuqi.mini.springframework.beans.factory.config.BeanDefinition;
+
+//负责 管理 BeanDefinition,规定：容器“必须”能注册和获取 BeanDefinition，那么可以使用接口
+public interface BeanDefinitionRegistry {
+    void registerBeanDefinition(String beanName, BeanDefinition beanDefinition);
+
+}