+++ src/main/java/cn/yuqi/mini/springframework/core/io/io/Resource.java
@@ -1,0 +1,7 @@
+package cn.yuqi.mini.springframework.core.io.io;
+
+import java.io.InputStream;
+
+public interface Resource {
+    InputStream getInputStream();
+}