+++ src/main/java/cn/yuqi/mini/springframework/core/io/io/ResourceLoader.java
@@ -1,0 +1,5 @@
+package cn.yuqi.mini.springframework.core.io.io;
+
+public interface ResourceLoader {
+    Resource getResource(String location);
+}